Fix pagetable1 sweeping on 32bit
If no pool allocations have occurred before the first GC (e.g.
because MEMDEBUG was set), the pagetable1 that covers the entire
(32bit) memory region on 32 bit machines may not have been
initialized yet, while the corresponding sweep function was making
the assumption that the passed in pagetable was non-null.
Add the appropriate check for initialization.