Make a missing GC root slightly more obvious to the checker
Finding the missing GC root is really hard for the checker. The correct
error message is on the store to newargs, complaining that we are rooting
a value that may have been GC'ed. However, `newargs` gets moved around
and re-assigned all over that place in such a way that the analyzer has
a hard time tracking it (I've seen it be successful if that function
is cut down a bit, so it might also be a heuristics problem). By
introducing a fake use that is easier for the analyzer to reason about,
hopefully we can catch regressions to this particular (very hairy)
part of the code in the future.