Add an assertion that DataTypes are allocated suitably aligned
We rely on DataTypes being allocated aligned, such that we may steal
the low order bits for GC flags. However, this is sort of an implicit
assumption guaranteed by pool alignment. Make this assumption explicit
by adding an assert, such that somebody will notice when porting to new
platforms where this may not be true (yet - because the appropriate
support wasn't hooked up).