intfuncs: allow use of literal power for power-by-squaring
For some types, x^2 can have a faster implementation than x*x: e.g.
polynomials benefit from ~half the number of operations.

Of course, for other types, the opposite can also happen, but that case
is taken care of by literal_pow(...). Theoretically, we can rely on that
mechanic for deciding the fastest implementation.

For backwards compatibility, we can't make this change in julia 1.x:
user-defined types may have method definitions implementing one in terms
of the other. This might give method ambiguities or stack overflow.

With this change, we therefore only give type authors the _option_
of using a literal pow. They'd do that by defining

    ^(x::MyType, n::Integer) = power_by_squaring(x, n, square=x->x^2)
    literal_pow(::typeof(^), x::MyType, ::Val{2}) = @todo # <faster than x*x>

In the future in Julia 2.0, we could see if this is important enough to
re-wire these methods.