Add weights argument to sum
This will be consistent with the `weights` argument to be added to similar functions
in Statistics. Since that code is needed notably to computed the weighted `mean`
in Statistics, I figured it is better to make it public rather than only defining
internal helpers for `mean`.

This is based on code from StatsBase, of which one part needs to live in LinearAlgebra
since it depends on BLAS and `dot`.