Clean up return->call address translation for StackTraces.lookup / jl_unw_stepn
Translate return->call address internally in jl_unw_stepn so that users
never have to deal with this subtlety when calling StackTraces.lookup.

Also rearrange the jl_unw_stepn API a little so that interpreter stack
traces can always be fit in to bt_data when calling jl_unw_stepn
incrementally for deep stacks + add test cases for that (Fixes #29695).
For the tests, clean up printing of CodeInfo objects wrapped in
InterpreterIP.