Profile: print paths more concisely
Heuristically decide on a short path representation while printing.
Look for a Project.toml file in a parent directly, take the name from
it, and substitute that for the prefix of the absolute path.

For relative paths, determine whether they are relative to the root of Base,
or not actually paths (such as REPL[1]).