Document customizable binary operators
I learned indirectly through comments on a PR that \otimes and \oplus and \odot can be extended as binary operators yet I cannot find this fact anywhere in the documentation.  Search the docs for `oplus` yields nothing.
Somewhere there must be a list of which characters can be extended this way and it should be documented in the manual.
This PR is an attempt to get that ball rolling!