Optimize partitioned views of CartesianIndices
And also recompute ranges instead of using views for partitions of ranges.

Since `Iterators.partition` is so handy for dividing up iteration spaces, it makes sense to optimize this as much as possible. While it is enherently a "linear" operation, it is a batched linear operation that allows us to skip doing all the effective ind2sub work on every single iteration.