Add test-revise-* Makefile targets (#33682)
These targets use Revise to load any modifications to base into the
process before running tests. This is very handy when editing Base
and wanting to run the appropriate tests as a sanity check but
without waiting 3 minutes for a whole system image rebuild. Of course
it was previously possible to do this through the REPL by doing the
3-4 steps manually, but I was generally too lazy too do that, so
here's the automatic option.