WIP: Only compute effect-free once
At the moment, whenever a statement becomes unused the compiler
tries to determine whether that statement is effect free and
may thus be removed. This is wasteful, because effect-freeness
should be computable only from type information. This PR moves
that computation to the start of the IRCode construction phase
and caches the result in the flags array. Ideally, I'd like to
just propagate these kinds of flags interprocedurally during
inference, but I'd like to do this one step at a time.