REPL: fix bug when all the input buffer can't be printed
When the input buffer contained too many lines, it was
still possible to edit it, but blindly for the n invisible
lines: indeed, the first n lines where never shown even
when the cursor was moved in their region.
Fix that, by centering the cursor in the available space
when all the lines can't be shown at the same time.
Other strategies are possible, some nicer, but this is
probably the simplest, by far, as it is "stateless": it
doesn't need to remember what happened during the last
refresh or to know what event triggered the refresh.