codegen: optimize returning an argument
It's not unusual to have code that simply returns an argument.  When
used in generic code though, that currently might involve copying the
value to a new box. We sometimes define functions like
`identity(nospecialize x) = x` to work around that, (which is preferable
anyways for other reasons), but it's better for codegen to also just be
aware of this pattern.