Clearer API for custom log levels
Add AbstractLogLevel as the supertype of all log level types.

Introduce "log level severity" including a `severity` function which
maps any user-defined log level to an integer log level severity. This
more cleanly separates out the ordinal notion of log severity from the
categorical aspects of the log level.

Log levels can be ordered by comparing their severities. This gives a
certian type of pre-order which is reflexive, transitive and total, but
not antisymmetric. This is the same as the ordering of complex numbers
by their magnitude.

Change to using `print` rather than `show` for formatting log levels for
display. This seems to resolve the old conundrum of "do we print Warn as
it appears in the program as "Warn" or in human-friendly form as
"Warning""?

The early (global) decision about log level filtering within the macro
is now lowered to `shouldlog(level)` which is more natural and more
customizable.