Fix #32980
This was introduced in: e2821f594c5e9b2ec0fa1603f4e90628d2e40cd7

Package managers track which package owns a file, and if the
icon-theme.cache is owned by Julia, the package manager will get
confused when that file already exists on the system or when another
process overwrites it.

It's the DE's job to update that cache.