sigwait: ignore errno except on Darwin
I am observing this value get set to EINTR sometimes on Linux. It is not
generally supposed to do that (though not prohibited either). That is
causing us to sometimes drop signals (usually SIGINT). Since we just
need this code for Darwin (ref #17626, #22041, #20174), we can remove
this from everywhere else.