runtests: Print currently running tests when asked
With this patch, pressing '?' while the test are running will print
which tests are still running. Of course, you could just try to look
through the log to try to figure out which tests have started but not
yet finished, but that's pretty annoying to do. Being able to just ask
is much easier for the impatient developer.