Use unchecked conversion in bitcounting operations. NFC. (#34398)
LLVM's intrinsics for bitcounting return an integer of the same size as
the input integer. On the julia side, we then trunc back to `Int` (since
we're guaranteed that the answers will always be at most `log2(x)`). However,
the code was calling for a checked conversion, rather than an unchecked one.
This doesn't matter too much as LLVM does know about the result range of these
intrinsics, but it made me sad to generate all these error checks in the first
place, so use the unchecked variants instead.