Remove legacy doc content
Remove the paragraphs saying "Immutable composite types with no fields are singletons".

Previously, `MutableNoFields() === MutableNoFields()`, but this was changed in #25854. When that got changed, the existing docs was tweaked to say that the struct had to be immutable(#25862) . However, there is no longer anything specific about it having no fields. The docs about singletons can thus just be removed since they are not special anymore.