Remove checked conversions in two more cases. NFC. (#34406)
Like #34398, but in a different place. I also realized there was a good reason to
do this beside personal preference: Once we start tracking nothrow more closely (to
optimize try/catch, etc.), we we'll want to make sure not to emit error paths that the
julia optimizer itself can't eliminate.