Force `clang`-based platforms to choose a cxx ABI
MacOS and FreeBSD are `clang` based, and so don't typically care about
our cxxstring ABI versioning, but we still build them because it is
_possible_ to use `g++` with `libstdc++` on MacOS and FreeBSD, so we
still need to build the two versions.  For the typical case, however,
the julia that is built by `clang` won't care which one you use, and so
this ensures that we just download the `cxx11` one by default.