Small tweak to embedding.md (#34422)
Adds -rpath to the link line which was suggested in #28982 and removes the line about JULIA_BINDIR, which isn't necessary anymore, since we autodetect the directory. Also remove the reference to libstdc++. If you built julia yourself, it's definitely compatible with the system one. If not, we now build julia binary distributions with wide compatibility with different libstdc++ versions, so the system one should be fine.