improve error message for using **
based on https://discourse.julialang.org/t/syntax-use-instead-of-and-syntax-is-not-a-unary-operator/34258/5, it seems that while well intentioned, this error message might be causing more confusion than it fixes. I took a stab at addressing it, but I think possibly the better change would be to parse `**` as an operator, and then we could define specific methods that threw better error messages. Specifically, `a**b` should have the error, but if we see `**a`, we should instead be telling people to use splatting, as that is more likely what they were trying to do.