Improve error message for trailing colon in certain situations
The expression
```
function test_function(inputs):
    ...
end
```
currently gives the following syntax error
```
ERROR: syntax: space not allowed after ":" used for quoting
```
This can be a bit confusing for users who do not know
about quoting and do not intend it to be used as such.
Since this kind of mistake is common for users coming from
Python, try to give an improved error message.