windows: Add script for automatically creating minidumps on timeout
This adds an autodump script in `contrib/autodump.jl`. This is intended
to be used on CI as `julia autodump.jl julia runtests.jl` and will,
on timeout automatically generate a memory dump of all the running processes
for later offline analysis. The idea here is to try to debug the hangs
that have been plaguing windows CI lately.