Convert range type in `reduced_index`
The `reduced_indices` and `reduced_indices0` methods sometimes assert
that the return axes type is the same as the input.
Consequently, the implementation of `reduced_index` had better
return a range of the same type as the input.

This corrects the error in
https://github.com/JuliaArrays/OffsetArrays.jl/issues/92.
I'll put a workaround in OffsetArrays.jl too.

(cherry picked from commit eda4131859229c7d9de4461e154a73bb89a6b783)