Add patch for 31156
This imports the patch I put up in https://reviews.llvm.org/D75072
and should fix #31156. We should probably hold off on merging this
for a few days while upstream review is ongoing. In the meantime,
this branch should be convenient to try. Make sure to remember to build
LLVM from source, not BB.