improve type lookup performance
Currently, this uses an expensive tree lookup and requires a lock.
Switching this to a hash function instead of a comparison will allow us
to make this much faster, especially for insertion.

and makes the actual changes to improve type lookup performance

And eliminates the linearcache

This Folds together the computation of the hash with the determination if it
has one, saving us at least one entire pass over the input types.

This memoizes the hash function early, so we never need to recompute it

fix possible issues with type_hash correctness

redundant with cache, though the probing design of these dict
implementions (IdDict, htable, cache) is broken and needs to be fixed (cf PR 15176)

eliminates the uid field, rebuild TypeMapLevel array as simple IdDict
(jl_unwrap_unionall is expensive, so an honest-to-goodness IdDict is
actually much faster here, and about the same memory cost, as a TypeSet)

make Method.specializations into a TypeSet